move flask app from examples and add all pipelines to web app
