+++ backend_jobs/ingestion_pipeline/main.py
@@ -131,7 +131,7 @@ def run(input_provider_name, input_provider_args=None, output_name=None, run_loc
 
             num_of_pages = image_provider.get_num_of_pages()
             create_batch = pipeline | 'create' >> \
-                apache_beam.Create([i for i in range(1, int(num_of_pages)+1)])
+                apache_beam.Create([i for i in range(1, int(20)+1)])
             images = create_batch | 'call API' >> \
                 apache_beam.ParDo(image_provider.get_images)
             extracted_elements = images | 'extract attributes' >> \
