+++ flask_app/templates/all_pipelines.html
