+++ flask_app/templates/ingestion_removal.html
