+++ flask_app/templates/recognition_removal.html
